requirements.txt
alembic==1.11.1
anyio==3.7.1
arrow==1.2.3
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.99.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
jose==1.0.0
make==0.1.6.post2
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.11
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
